+++ src/django_perf_rec/sql.py
@@ -81,6 +81,21 @@ def sql_recursively_simplify(node, hide_columns=True):
         middle = [Token(tokens.Punctuation, " ... ")]
         node.tokens = node.tokens[: i_set + 1] + middle + node.tokens[i_where:]
 
+    # Ensure IN clauses with simple value in always simplify to "..."
+    if node.tokens[0].value == "WHERE":
+        in_token_indices = (i for i, t in enumerate(node.tokens) if t.value == "IN")
+        for in_token_index in in_token_indices:
+            parenthesis = next(
+                t
+                for t in node.tokens[in_token_index + 1 :]
+                if isinstance(t, Parenthesis)
+            )
+            if all(
+                getattr(t, "ttype", "") in sql_deleteable_tokens
+                for t in parenthesis.tokens[1:-1]
+            ):
+                parenthesis.tokens[1:-1] = [Token(tokens.Punctuation, "...")]
+
     # Erase the names of savepoints since they are non-deteriministic
     if hasattr(node, "tokens"):
         # SAVEPOINT x